Created UserModel::create()
This method will create a user after encoding its password and return
a User object with the encoded password and user ID.

I also switched mockery version to dev-master in order to get a fix
allowing Demeter chains with parameters mocking.
The Silex/Symfony password encoder is at the end of a Demeter chain.